# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FoodTinder API",
        "description": "API do logowania, swipe’ów i rekomendacji dań. Po prostu Tinder dla jedzenia!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/recommended-dishes": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Rekomendacje dla użytkowników",
                "description": "Zwraca listę rekomendowanych dań dla aktualnego użytkownika.",
                "operationId": "recommendation",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maksymalna liczba pozycji",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "",
                            "maximum": 20,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DishWithScore"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 101,
                                        "name": "Spaghetti Carbonara",
                                        "category_id": 1,
                                        "cuisine_id": 3,
                                        "flavour_id": 5,
                                        "image_url": "spaghetti.jpg",
                                        "image_url_full": "http://localhost/storage/spaghetti.jpg",
                                        "description": "Classic",
                                        "match_score": 0.93
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Dish": {
                "required": [
                    "id",
                    "name",
                    "category_id",
                    "cuisine_id",
                    "flavour_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "name": {
                        "type": "string",
                        "example": "Spaghetti Carbonara"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cuisine_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "flavour_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "image_url": {
                        "type": "string",
                        "example": "spaghetti.jpg",
                        "nullable": true
                    },
                    "image_url_full": {
                        "type": "string",
                        "example": "http://localhost/storage/spaghetti.jpg"
                    },
                    "description": {
                        "type": "string",
                        "example": "Classic Roman pasta.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DishWithScore": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Dish"
                    },
                    {
                        "properties": {
                            "match_score": {
                                "type": "number",
                                "format": "float",
                                "example": 0.93
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}